requirements.txt
Django>=3.1.7,<3.2.0
djangorestframework>=3.12.2,<3.13.0

python-decouple>=3.4,<3.5
djangorestframework-simplejwt>=4.6.0,<4.7.0
django-filter>=2.4.0,<2.4.1
